--- a/pkg/kv/kvserver/batcheval/cmd_refresh_range.go
+++ b/pkg/kv/kvserver/batcheval/cmd_refresh_range.go
@@ -89,7 +89,7 @@ func refreshRange(
 			break
 		}
 
-		key := iter.Key()
+		key := iter.UnsafeKey().Clone()
 		if !key.IsValue() {
 			// Found an intent. Check whether it is owned by this transaction.
 			// If so, proceed with iteration. Otherwise, return an error.