@erikgrinaker 2022-06-16
> Does it make sense to have a test that checks if other KeyTypes work? PointsOnly is likely covered by existing tests since we pass the value down, but RangesOnly might be interesting.

Yeah, good call -- added some tests for `PointsOnly` and `RangesOnly`.